% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_parameter.R
\name{all_param}
\alias{all_param}
\title{parameters}
\arguments{
\item{param_atmos_thr_Ts}{<-1, 3> (Cel) threshold air temperature that snow, parameter for \code{\link[=atmosSnow_ThresholdT]{atmosSnow_ThresholdT()}}}

\item{param_atmos_ubc_A0FORM}{<0.01, 3> (Cel) threshold air temperature that snow, it can not equal or small than 0, parameter for \code{\link[=atmosSnow_UBC]{atmosSnow_UBC()}}}

\item{param_baseflow_grf_gamma}{<2, 7> exponential parameter for \code{\link[=baseflow_GR4Jfix]{baseflow_GR4Jfix()}}}

\item{param_baseflow_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=baseflow_SupplyRatio]{baseflow_SupplyRatio()}}}

\item{param_baseflow_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=baseflow_SupplyPow]{baseflow_SupplyPow()}}}

\item{param_baseflow_sup_gamma}{<0, 1> exponential parameter for \code{\link[=baseflow_SupplyPow]{baseflow_SupplyPow()}}}

\item{param_baseflow_map_gamma}{<0.1, 5> exponential parameter for \code{\link[=baseflow_MaxPow]{baseflow_MaxPow()}}}

\item{param_baseflow_thp_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_baseflow_thp_gamma}{<0.1, 5> exponential parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_baseflow_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_baseflow_arn_k}{<0.1, 1> exponential parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_capirise_acr_k}{<0.01, 1> coefficient parameter \code{\link[=capirise_AcceptRatio]{capirise_AcceptRatio()}}}

\item{param_capirise_acp_k}{<0.01, 1> coefficient parameter for \code{\link[=capirise_AcceptPow]{capirise_AcceptPow()}}}

\item{param_capirise_acp_gamma}{<0.01, 1> exponential parameter for \code{\link[=capirise_AcceptPow]{capirise_AcceptPow()}}}

\item{param_confluen_kel_k}{<1, 4> parameter for\code{\link[=confluenIUH_Kelly]{confluenIUH_Kelly()}}}

\item{param_confluen_nas_n}{<1, 8> parameter for\code{\link[=confluenIUH_Nash]{confluenIUH_Nash()}}}

\item{param_evatrans_tur_k}{<0.6, 1> parameter for \code{\link[=evatransPotential_TurcWendling]{evatransPotential_TurcWendling()}}, higher value when closer to the sea}

\item{param_evatrans_sur_k}{<0.1, 1> parameter for \code{\link[=evatransActual_SupplyRatio]{evatransActual_SupplyRatio()}}, ratio of potential ET, that is estimated as actually ET}

\item{param_evatrans_sup_k}{<0.1, 1> parameter for \code{\link[=evatransActual_SupplyPow]{evatransActual_SupplyPow()}}, ratio of this method}

\item{param_evatrans_sup_gamma}{<1, 5> parameter for \code{\link[=evatransActual_SupplyPow]{evatransActual_SupplyPow()}}, exponent of this method}

\item{param_evatrans_vic_gamma}{<0.2, 5> parameter for \code{\link[=evatransActual_VIC]{evatransActual_VIC()}}}

\item{param_evatrans_ubc_gamma}{<0.5, 2> parameter for \code{\link[=evatransActual_UBC]{evatransActual_UBC()}}}

\item{param_evatrans_lia_gamma}{<0.4, 1> parameter for \code{\link[=evatransActual_LiangLand]{evatransActual_LiangLand()}}}

\item{param_evatrans_lia_B}{<0.01, 3> parameter for \code{\link[=evatransActual_LiangSoil]{evatransActual_LiangSoil()}}}

\item{param_infilt_ubc_P0AGEN}{<0.1, 4> coefficient parameter for \code{\link[=infilt_UBC]{infilt_UBC()}}}

\item{param_infilt_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_SupplyRatio]{infilt_SupplyRatio()}}}

\item{param_infilt_acr_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_AcceptRatio]{infilt_AcceptRatio()}}}

\item{param_infilt_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_SupplyPow]{infilt_SupplyPow()}}}

\item{param_infilt_sup_gamma}{<0, 1> parameters for \code{\link[=infilt_SupplyPow]{infilt_SupplyPow()}}}

\item{param_infilt_acp_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_AcceptPow]{infilt_AcceptPow()}}}

\item{param_infilt_acp_gamma}{<0.001, 5> parameters for \code{\link[=infilt_AcceptPow]{infilt_AcceptPow()}}}

\item{param_infilt_hbv_beta}{<0.001, 5> parameters for \code{\link[=infilt_HBV]{infilt_HBV()}}}

\item{param_infilt_xaj_B}{<0.01, 3> parameters for \code{\link[=infilt_XAJ]{infilt_XAJ()}}}

\item{param_infilt_vic_B}{<0.01, 3> parameters for \code{\link[=infilt_VIC]{infilt_VIC()}}}

\item{param_inteflow_grf_k}{<0.01, 1> coefficient parameter for \code{\link[=inteflow_GR4Jfix]{inteflow_GR4Jfix()}}}

\item{param_inteflow_grf_gamma}{<2, 7> exponential parameter for \code{\link[=baseflow_GR4Jfix]{baseflow_GR4Jfix()}}}

\item{param_inteflow_map_gamma}{<0.1, 5> exponential parameter for \code{\link[=inteflow_MaxPow]{inteflow_MaxPow()}}}

\item{param_inteflow_thp_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=inteflow_ThreshPow]{inteflow_ThreshPow()}}}

\item{param_inteflow_thp_gamma}{<0.1, 5> exponential parameter for \code{\link[=inteflow_ThreshPow]{inteflow_ThreshPow()}}}

\item{param_inteflow_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=inteflow_ThreshPow]{inteflow_ThreshPow()}}}

\item{param_inteflow_arn_k}{<0.1, 1> exponential parameter for \code{\link[=inteflow_ThreshPow]{inteflow_ThreshPow()}}}

\item{param_inteflow_sp0_k}{<0.01, 1> coefficient parameter for \code{\link[=inteflow_SupplyPow0]{inteflow_SupplyPow0()}}}

\item{param_inteflow_sp0_gamma}{<0, 1> exponential parameter for \code{\link[=inteflow_SupplyPow0]{inteflow_SupplyPow0()}}}

\item{param_inteflow_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=inteflow_SupplyPow]{inteflow_SupplyPow()}}}

\item{param_inteflow_sup_gamma}{<0, 7> parameter for \code{\link[=inteflow_SupplyPow]{inteflow_SupplyPow()}}}

\item{param_inteflow_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=inteflow_SupplyRatio]{inteflow_SupplyRatio()}}}

\item{param_lateral_sup_k}{<-1, 1> coefficient parameter for \code{\link[=lateral_SupplyPow]{lateral_SupplyPow()}}}

\item{param_lateral_sup_gamma}{<0.01, 5> parameters for \code{\link[=lateral_SupplyPow]{lateral_SupplyPow()}}}

\item{param_lateral_sur_k}{<-2, 1> coefficient parameter for \code{\link[=lateral_SupplyRatio]{lateral_SupplyRatio()}}}

\item{param_lateral_grf_gamma}{<0.01, 5> parameter for \code{\link[=lateral_GR4Jfix]{lateral_GR4Jfix()}}}

\item{param_lateral_thp_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=lateral_ThreshPow]{lateral_ThreshPow()}}}

\item{param_lateral_thp_gamma}{<0.1, 5> exponential parameter for \code{\link[=lateral_ThreshPow]{lateral_ThreshPow()}}}

\item{param_lateral_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=lateral_ThreshPow]{lateral_ThreshPow()}}}

\item{param_lateral_arn_k}{<0.1, 1> exponential parameter for \code{\link[=lateral_ThreshPow]{lateral_ThreshPow()}}}

\item{param_percola_grf_k}{<0.01, 1> coefficient parameter for \code{\link[=percola_GR4Jfix]{percola_GR4Jfix()}}}

\item{param_percola_map_gamma}{<0.1, 5> exponential parameter for \code{\link[=percola_MaxPow]{percola_MaxPow()}}}

\item{param_percola_thp_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_thp_gamma}{<0.1, 5> exponential parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_arn_k}{<0.1, 1> exponential parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=percola_SupplyPow]{percola_SupplyPow()}}}

\item{param_percola_sup_gamma}{<0, 7> parameter for \code{\link[=percola_SupplyPow]{percola_SupplyPow()}}}

\item{param_percola_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=percola_SupplyRatio]{percola_SupplyRatio()}}}

\item{param_snow_kus_fE}{<0.0005, 0.003> (mm/m2/MJ) snow melt temperature parameter for \code{\link[=snowMelt_Factor]{snowMelt_Factor()}}}

\item{param_snow_kus_fT}{<0.05, 1> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for \code{\link[=snowMelt_Factor]{snowMelt_Factor()}}}

\item{param_snow_fac_Tmelt}{<0, 3> (Cel) snow melt temperature parameter for \code{\link[=snowMelt_Factor]{snowMelt_Factor()}}}

\item{param_snow_fac_f}{<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for \code{\link[=snowMelt_Factor]{snowMelt_Factor()}}}
}
\description{
all of the parameters in the EDCHM
}
