% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_variable.R
\name{all_vari}
\alias{all_vari}
\title{variable define}
\arguments{
\item{time_step_h}{<1, 24> (h) time step in hour}

\item{time_dayOfYear_}{<1, 366> (d) the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)}

\item{atmos_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{atmos_rain_mm}{(mm/m2/TS) precipitation in rain form}

\item{atmos_snow_mm}{(mm/m2/TS) precipitation in snow form}

\item{atmos_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{atmos_temperatureMax_Cel}{(Cel) the maximal air temperature in the time phase}

\item{atmos_temperatureMin_Cel}{(Cel) the minimal air temperature in the time phase}

\item{atmos_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{atmos_netRadiat_MJ}{(MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation}

\item{atmos_relativeHumidity_1}{(0, 1) relative humidity}

\item{atmos_vaporPress_hPa}{(hPa) actual vapour pressure, can be calculated by \code{\link[=atmos_VaporPress]{atmos_VaporPress()}}}

\item{atmos_saturatVaporPress_hPa}{(hPa) saturation vapour pressure at \code{atmos_temperature_Cel}, can be calculated by \code{\link[=atmos_SaturatVaporPress]{atmos_SaturatVaporPress()}}}

\item{atmos_windSpeed_m_s}{(m/s) measured wind speed at z m above ground surface}

\item{atmos_windMeasureHeight_m}{(m) height of measurement above ground surface}

\item{atmos_windSpeed2m_m_s}{(m/s) wind speed at 2 m above ground surface}

\item{atmos_potentialEvatrans_mm}{(mm/m2/TS) \strong{potential / reference} evapotranspiration}

\item{land_albedo_1}{<0, 1> albedo of the region}

\item{land_latitude_Degree}{(degree) average latitude}

\item{land_elevation_m}{(m) average elevation}

\item{land_water_mm}{(mm/m2) \strong{pounded water} volume in \code{landLy} and there is no limit, different than \code{land_interceptWater_mm}}

\item{land_interceptWater_mm}{(mm/m2) initial water volume that can be \strong{intercepted}}

\item{land_interceptCapacity_mm}{(mm/m2) average intercept Capacity (maximal storage capacity)}

\item{land_actualEvatrans_mm}{(mm/m2/TS) \strong{actual} evapotranspiration from \code{landLy}}

\item{land_infiltrat_mm}{(mm/m2/TS) infiltration from \code{landLy} to \code{soilLy}}

\item{snow_ice_mm}{(mm/m2) water equivalent of \strong{ice} in snowpack}

\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{soil_actualEvatrans_mm}{(mm/m2/TS) \strong{actual} evapotranspiration from \code{soilLy}}

\item{ground_water_mm}{(mm/m2/TS) water volume in \code{groundLy}}

\item{ground_capacity_mm}{(mm/m2) water storage capacity in \code{groundLy}}

\item{ground_lateral_mm}{(mm/m2/TS) lateral flow, exchange with outside region. It can be \strong{NEGATIV}}

\item{ground_capillarise_mm}{(mm/m2/TS) capillary rise from \code{groundLy} to \code{soilLy}}

\item{ground_lateralPotential_mm}{(mm/m2/TS) potential lateral flow, that will as parameter evaluated}

\item{confluen_inputWater_mm, land_runoff_mm, ground_baseflow_mm}{(mm/m2) input water volum in every routeline}

\item{confluen_iuh_1, confluen_iuhLand_1, confluen_iuhGround_1}{(vector of num, sume() = 1) the ratio in every timestep, can be calculated by \code{\link[=confluenIUH_GR4J1]{confluenIUH_GR4J1()}}, \code{\link[=confluenIUH_GR4J2]{confluenIUH_GR4J2()}}}

\item{confluen_responseTime_TS, confluen_concentratTime_TS}{(TS) response or concentration time in every routeline}

\item{water_mm}{(mm/m2/TS) water volume in \code{soilLy} or interceptof \code{landLy}, same as \code{land_interceptWater_mm} or \code{soil_water_mm}}

\item{capacity_mm}{(mm/m2) water storage capacity in \code{soilLy} or interceptof \code{landLy}, same as \code{land_interceptCapacity_mm} or \code{soil_capacity_mm}}
}
\description{
In this topic all of the variable in the EDCHM will be defined.
}
