% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{atmosSnow}
\alias{atmosSnow}
\alias{atmosSnow_ThresholdT}
\alias{atmosSnow_UBC}
\title{caculate \strong{snowfall}}
\usage{
atmosSnow_ThresholdT(
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  param_atmos_thr_Ts
)

atmosSnow_UBC(
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  param_atmos_ubc_A0FORM
)
}
\arguments{
\item{atmos_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{atmos_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{param_atmos_thr_Ts}{(Cel) parameters for \code{\link[=atmosSnow_ThresholdT]{atmosSnow_ThresholdT()}}}

\item{param_atmos_ubc_A0FORM}{(0, 3 Cel) parameters for \code{\link[=atmosSnow_UBC]{atmosSnow_UBC()}}}
}
\value{
atmos_snow_mm (mm/m2/TS) snowfall volum
}
\description{
caculate \strong{snowfall}
}
