% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseflow}
\alias{baseflow}
\alias{baseflow_GR4J}
\alias{baseflow_GR4Jfix}
\alias{baseflow_SupplyPow}
\alias{baseflow_SupplyRatio}
\alias{baseflow_MaxPow}
\alias{baseflow_ThreshPow}
\alias{baseflow_Arno}
\title{\strong{baseflow}}
\usage{
baseflow_GR4J(ground_water_mm, ground_capacity_mm)

baseflow_GR4Jfix(ground_water_mm, ground_capacity_mm, param_baseflow_grf_gamma)

baseflow_SupplyPow(
  ground_water_mm,
  param_baseflow_sup_k,
  param_baseflow_sup_gamma
)

baseflow_SupplyRatio(ground_water_mm, param_baseflow_sur_k)

baseflow_MaxPow(
  ground_water_mm,
  ground_capacity_mm,
  ground_potentialBaseflow_mm,
  param_baseflow_map_gamma
)

baseflow_ThreshPow(
  ground_water_mm,
  ground_capacity_mm,
  ground_potentialBaseflow_mm,
  param_baseflow_thp_thresh,
  param_baseflow_thp_gamma
)

baseflow_Arno(
  ground_water_mm,
  ground_capacity_mm,
  ground_potentialBaseflow_mm,
  param_baseflow_arn_thresh,
  param_baseflow_arn_k
)
}
\arguments{
\item{ground_water_mm}{(mm/m2/TS) water volume in \code{groundLy}}

\item{ground_capacity_mm}{(mm/m2) water storage capacity in \code{groundLy}}

\item{param_baseflow_grf_gamma}{<2, 7> exponential parameter for \code{\link[=baseflow_GR4Jfix]{baseflow_GR4Jfix()}}}

\item{param_baseflow_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=baseflow_SupplyPow]{baseflow_SupplyPow()}}}

\item{param_baseflow_sup_gamma}{<0, 1> exponential parameter for \code{\link[=baseflow_SupplyPow]{baseflow_SupplyPow()}}}

\item{param_baseflow_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=baseflow_SupplyRatio]{baseflow_SupplyRatio()}}}

\item{ground_potentialBaseflow_mm}{<0.01, 7> (mm/m2/TS) \strong{potential} baseflow}

\item{param_baseflow_map_gamma}{<0.1, 5> exponential parameter for \code{\link[=baseflow_MaxPow]{baseflow_MaxPow()}}}

\item{param_baseflow_thp_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_baseflow_thp_gamma}{<0.1, 5> exponential parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_baseflow_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}

\item{param_baseflow_arn_k}{<0.1, 1> exponential parameter for \code{\link[=baseflow_ThreshPow]{baseflow_ThreshPow()}}}
}
\value{
ground_baseflow_mm (mm/m2/TS)
}
\description{
\strong{baseflow}
}
