% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali.R
\name{cali}
\alias{cali}
\alias{cali_DDS}
\alias{cali_UVS}
\title{Calibrate algorithms}
\usage{
cali_DDS(fitness, x_Min, x_Max, x_Init = NA, max_iter = 100, r = 0.2, ...)

cali_UVS(
  fitness,
  x_Min,
  x_Max,
  x_Init = NA,
  max_n_loop = 10,
  n_1_search = 8,
  ...
)
}
\arguments{
\item{fitness}{fitness function}

\item{x_Min, x_Max, x_Init}{minimal, maximal initial parameter}

\item{max_iter}{maximal number of iteration}

\item{r}{parameter for algorithm}

\item{...}{other parameters}

\item{max_n_loop}{maximal number of loops}

\item{n_1_search}{maximal number in one loop}
}
\description{
Calibrate algorithms
}
\section{DDS \insertCite{DDS_Tolson_2007}{EDCHM}}{
dynamically dimensioned search (DDS) calibration algorithm
}

\section{UVS}{
Univariable search algorithmus
}

