% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confluen}
\alias{confluen}
\alias{confluen_IUH}
\alias{confluen_IUH2S}
\alias{confluen_IUH3S}
\title{\strong{confluence}}
\usage{
confluen_IUH(confluen_inputWater_mm, confluen_iuh_1)

confluen_IUH2S(
  land_runoff_mm,
  ground_baseflow_mm,
  confluen_iuhLand_1,
  confluen_iuhGround_1
)

confluen_IUH3S(
  land_runoff_mm,
  soil_interflow_mm,
  ground_baseflow_mm,
  confluen_iuhLand_1,
  confluen_iuhSoil_1,
  confluen_iuhGround_1
)
}
\arguments{
\item{confluen_inputWater_mm, land_runoff_mm, ground_baseflow_mm}{(mm/m2) input water volum in every routeline}

\item{confluen_iuh_1, confluen_iuhLand_1, confluen_iuhSoil_1, confluen_iuhGround_1}{(vector of num, sume() = 1) the ratio in every timestep, can be calculated by \code{\link[=confluenIUH_GR4J1]{confluenIUH_GR4J1()}}, \code{\link[=confluenIUH_GR4J2]{confluenIUH_GR4J2()}}}

\item{soil_interflow_mm}{(mm/m2/TS) subsurface flow directly to the river}
}
\value{
confluenced water (mm/m2)
}
\description{
\loadmathjax

In hydrological modeling, routing (named as \link{confluen} in EDCHM) refers to the process of simulating the movement of water through a river network or other drainage system.
It allows the model to predict the flow of water in rivers and streams.
In hydrological models, routing is typically performed using mathematical algorithms that account for the physical properties of the river network,
such as its geometry, roughness, and discharge capacity.
The parameters that govern routing, such as flow velocity and channel roughness,
can have a significant impact on the accuracy of the model.

\code{confluence} is a calculation function that causes water to flow into the gauge point.
\itemize{
\item \code{IUH}: IUH (Instant Unit Hydrograph) with one watercourse,
\item \code{IUH2S}: IUH with two water sources, each with a different IUH vector,
\item \code{IUH3S}: IUH with three water sources, each with a different IUH vector.
}

Under the concept of the conceptual HM, the water flux to the water flow will be calculated using the confluence process.
This process does not calculate the water balance, but rather the time-varying nature of the water flow.
The "Instant Unit Hydrograph" method is the most effective way to deal with time-varying flows.
In the first stage, only \link{confluenIUH} will be supported.

So we can give the function:

\mjsdeqn{Q = f_{confluen}(F, u)}

where
\itemize{
\item \mjseqn{Q} is stream flow, but still in mm/TS not m3/TS or m3/S
\item \mjseqn{F} is flux that will into river conflen, e.g.\code{land_runoff_mm}, \code{soil_interflow_mm} or \code{ground_baseflow_mm}
\item \mjseqn{u} is Instant Unit Hydrograph series
}
}
\references{
\insertAllCited{}
}
