% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confluenIUH}
\alias{confluenIUH}
\alias{confluenIUH_GR4J1}
\alias{confluenIUH_GR4J2}
\title{create \strong{IUH} (Instant Unit Graphy)}
\usage{
confluenIUH_GR4J1(confluen_resposeTime_TS)

confluenIUH_GR4J2(confluen_resposeTime_TS)
}
\arguments{
\item{confluen_resposeTime_TS}{(TS) respose time in every routeline}
}
\value{
IUH (list of num vector)
}
\description{
create \strong{IUH} (Instant Unit Graphy)
}
