% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDCHM-package.R
\docType{package}
\name{EDCHM-package}
\alias{EDCHM-package}
\title{basic concept of EDCHM}
\description{
EDCHM is a conceptional hydrologist modelling framework,
with that we can easy assembly a new hydrologist model with difficult process-method (modular).
}
\section{Layer and Boundary}{
Under EDCHM the \strong{Layer} is defined layer with vertical space, in that have some common characters.
And the Layer always have top and bottom two \strong{Boundary},
one Boundary in the same time the top Boundary of one Layer but also the bottom Layer of other Layer, that on the above.

So for EDCHM there are five basic Layers (now one additional snowLy) and six Boundaries:
\itemize{
\item Layer:
\itemize{
\item atmosLy
\item landLy (snowLy)
\item soilLy
\item groundLy
}
\item Boundary:
\itemize{
\item topBd
\item atmosBd
\item landBd
\item groundBd
\item btmBd
}
}
}

\section{Variable define}{
All the names of the variables make up of three parties: \strong{group-name}, \strong{physical-name} and \strong{variable-units}
and split with underline \verb{_}.
\itemize{
\item group-name: the Layer or Boundary name, which is defined in EDCHM.
Sometimes it can also be the Process name or \code{time}, one important dimension.
But the group-name is no longer than 8 and all small letters.
\item physical-name: the physical variable name, this part will with \emph{camelCase} to combine more words.
\item variable-units: the physical units of the variable, in order to simplify the program,
all the time dependent variable will use the \code{TS} as the time units, it will be defined by the model.
But the time units part will not appear in the name.
And also all the variable meant they are homogeneous in the area, so all the variable will not give the square meter (m2) in the name.
}

e.g. \strong{group_variableName_unit}
}

\section{Parameter define}{
The Parameter will defined in every function topic, but there will define the naming regulation:
Parameters make up with prefix \code{param}, Process name (sometimes same as Layer name),
abbreviation in threee small letters of the Method, and the Parameter name in the original.

e.g. \strong{param_processName_mtd_k}
}

\section{Process}{
There are now 10 Process available:
\itemize{
\item atmos: caculate the basic variable (data) in the \code{atmosLy}, sometimes the variable komm directly from the original data
\itemize{
\item atmosSnow: rain and snoe to split
}
\item evatrans: evapotranspiration
\itemize{
\item evatransPotential: potential ET calculate
\item evatransActual: actual ET calculate
\item evatransLand: actual ET in \code{landLy} (evaporation from interception)
\item evatransSoil: actual ET in \code{soilLy} (transpiration from root zone + evaporation from top soil)
}
\item snow: snow
\itemize{
\item snowMelt:
}
\item intercep: interception
\item infilt: infiltration \code{landLy} to \code{soilLy}
\item capirise: capital rise, \code{groundLy} to \code{soilLy}
\item percola: percolation, \code{soilLy} to \code{groundLy}
\item baseflow: interception, \code{groundLy}
\item lateral: lateral flow, exchange with outside region, \code{groundLy}
\item confluen: confluence to river
\itemize{
\item confluenIUH: IUH to calculate
}
}
}

