% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evatransActual}
\alias{evatransActual}
\alias{evatransActual_FestRatio}
\alias{evatransActual_Supply}
\alias{evatransActual_SupplyRatio}
\alias{evatransActual_SupplyPow}
\alias{evatransActual_VIC}
\alias{evatransActual_GR4J}
\alias{evatransActual_UBC}
\alias{evatransLand_Liang}
\alias{evatransSoil_Liang}
\title{\strong{actuall evapotranspiration}}
\usage{
evatransActual_FestRatio(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_fst_k
)

evatransActual_Supply(atmos_potentialEvatrans_mm, water_mm, capacity_mm)

evatransActual_SupplyRatio(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_sup_k
)

evatransActual_SupplyPow(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_pow_k,
  param_evatrans_pow_gamma
)

evatransActual_VIC(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_vic_gamma
)

evatransActual_GR4J(atmos_potentialEvatrans_mm, capacity_mm, water_mm)

evatransActual_UBC(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_infilt_ubc_P0EGEN
)

evatransLand_Liang(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_lia_gamma
)

evatransSoil_Liang(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_lia_B
)
}
\arguments{
\item{atmos_potentialEvatrans_mm}{(mm/m2) \strong{potential / reference} evapotranspiration}

\item{water_mm}{(mm/m2) water volum in \code{soilLy} or interceptof \code{landLy}}

\item{capacity_mm}{(mm/m2) water storage capacity in \code{soilLy} or interceptof \code{landLy}}

\item{param_evatrans_fst_k}{parameters for \code{\link[=evatransActual_FestRatio]{evatransActual_FestRatio()}}}

\item{param_evatrans_sup_k}{parameters for \code{\link[=evatransActual_SupplyRatio]{evatransActual_SupplyRatio()}}}

\item{param_evatrans_pow_k, param_evatrans_pow_gamma}{parameters for \code{\link[=evatransActual_SupplyPow]{evatransActual_SupplyPow()}}}

\item{param_evatrans_vic_gamma}{parameters for \code{\link[=evatransActual_VIC]{evatransActual_VIC()}}}

\item{param_infilt_ubc_P0EGEN}{parameters for \code{\link[=evatransActual_UBC]{evatransActual_UBC()}}}

\item{param_evatrans_lia_gamma}{parameters for \code{\link[=evatransLand_Liang]{evatransLand_Liang()}}}

\item{param_evatrans_lia_B}{parameters for \code{\link[=evatransSoil_Liang]{evatransSoil_Liang()}}}
}
\value{
actuall evapotranspiration (mm/m2) in one storage:
\itemize{
\item evaporation in interception (landLy)
\item transpiration in root
\item evaporation in soil (soilLy)
}
}
\description{
\strong{actuall evapotranspiration}
}
