% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evatransPotential}
\alias{evatransPotential}
\alias{evatransPotential_TurcWendling}
\alias{evatransPotential_Linacre}
\alias{evatransPotential_FAO56}
\title{\strong{potential evapotranspiration}}
\usage{
evatransPotential_TurcWendling(
  atmos_temperature_Cel,
  atmos_solarRadiat_MJ,
  time_step_h,
  param_evatrans_tur_k
)

evatransPotential_Linacre(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_relativeHumidity_1,
  land_latitude_Degree,
  land_elevation_m,
  land_albedo_1
)

evatransPotential_FAO56(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_vaporPress_hPa,
  atmos_saturatVaporPress_hPa,
  atmos_netRadiat_MJ,
  atmos_windSpeed2m_m_s,
  land_latitude_Degree,
  land_elevation_m
)
}
\arguments{
\item{atmos_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{atmos_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{time_step_h}{<1, 24> (h) time step in hour}

\item{param_evatrans_tur_k}{<0.6, 1> parameter for \code{\link[=evatransPotential_TurcWendling]{evatransPotential_TurcWendling()}}, higher value when closer to the sea}

\item{time_dayOfYear_}{<1, 366> the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)}

\item{atmos_relativeHumidity_1}{<0, 1> relative humidity}

\item{land_latitude_Degree}{(degree) average latitude}

\item{land_elevation_m}{(m) average elevation}

\item{land_albedo_1}{<0, 1> albedo of the region}

\item{atmos_vaporPress_hPa}{(hPa) actual vapour pressure, can be calculated by \code{\link[=atmos_VaporPress]{atmos_VaporPress()}}}

\item{atmos_saturatVaporPress_hPa}{(hPa) saturation vapour pressure at \code{atmos_temperature_Cel}, can be calculated by \code{\link[=atmos_SaturatVaporPress]{atmos_SaturatVaporPress()}}}

\item{atmos_netRadiat_MJ}{(MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation}

\item{atmos_windSpeed2m_m_s}{(m/s) wind speed at 2 m above ground surface}
}
\value{
potential evapotranspiration (mm/m2)
}
\description{
The concept potential ET estimate mainly the confluence from atmosphere,
when the water in the ET-area is always enough.
Actually we should also consider the charterers of the ET-area, there is free water area or vegetation or bar soil,
But we may don't have so many information and effective method, in order to get the truly potential ET.
So there is the simplified method: reference ET, it define the ET-area with some fest charterers, e.g. \code{\link[=evatransPotential_FAO56]{evatransPotential_FAO56()}}.
In this situation we should give some factor for different ET-area.
}
\details{
\itemize{
\item \strong{_TurcWendling}: consider only the radiation and temperature as the main factors.
\mjsdeqn{E_p = \frac{(100 R_s + 3.875 t_h k)\cdot(T + 22)}{150 (T + 123)}}
where
\itemize{
\item \mjseqn{E_p} is potential ET, \code{atmos_potentialEvatrans_mm}
\item \mjseqn{R_s} is solar radiation, \code{atmos_solarRadiat_MJ}
\item \mjseqn{t_h} is time step in hour, \code{time_step_h}
\item \mjseqn{T} is average air temperature, \code{atmos_temperature_Cel}
\item \mjseqn{k} is \code{param_evatrans_tur_k}
}
}

\itemize{
\item \strong{_Linacre}: consider only the temperature as the main factors.
\mjsdeqn{E_p = \frac{\frac{100(0.75 - \alpha)(T + 0.006 z)}{100 - \phi} + 15(T - T_d)}{80 - T}}
\mjsdeqn{T_d = T - 20 (1-H_R)}
where
\itemize{
\item \mjseqn{\alpha} is albedo, \code{land_albedo_1}
\item \mjseqn{z} is elevation, \code{land_elevation_m}
\item \mjseqn{T_d} is dewpoint temperature,
\item \mjseqn{H_R} is relative humidity, \code{atmos_relativeHumidity_1}
}
}

\itemize{
\item \strong{_FAO56}: consider not only radiation and temperature but also other variable like wind speed
as the main factors.
\mjsdeqn{E_p =\frac{0.408 \Delta\left(R_n - G\right)+\gamma \frac{900}{T+273} {u}_{2}\left({e}_{{s}}-{e}_{{a}}\right)}{\Delta+\gamma\left(1+0.34 {u}_{2}\right)}}
where
\itemize{
\item \mjseqn{\Delta} is slope vapour pressure curve (kPa °C-1)
\item \mjseqn{R_n} is net radiation, \code{atmos_netRadiat_MJ}
\item \mjseqn{G} is soil heat flux density
\item \mjseqn{u_2} is wind speed at 2 m height, \code{atmos_windSpeed2m_m_s}
\item \mjseqn{e_s} is saturation vapour pressure, \code{atmos_saturatVaporPress_hPa}
\item \mjseqn{e_a} is actual vapour pressure, \code{atmos_vaporPress_hPa}
\item \mjseqn{\gamma} is psychrometric constant
}
}
}
