% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intercep}
\alias{intercep}
\alias{intercep_Full}
\title{\strong{interception} water from land go into the soil.}
\usage{
intercep_Full(atmos_rain_mm, land_interceptWater_mm, land_interceptCapacity_mm)
}
\arguments{
\item{atmos_rain_mm}{(mm/m2) preciptation in rain form}

\item{land_interceptWater_mm}{(mm/m2) initial water volum that can be intercepted}

\item{land_interceptCapacity_mm}{(mm/m2) average intercept Capacity (maximal storage capacity)}
}
\value{
intercept_water_mm (mm/m2) intercepted water in this timestep
}
\description{
\strong{interception} water from land go into the soil.
}
