% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lateral}
\alias{lateral}
\alias{lateral_GR4J}
\alias{lateral_SupplyRatio}
\alias{lateral_SupplyPow}
\title{\strong{lateral flux}}
\usage{
lateral_GR4J(ground_water_mm, ground_capacity_mm, param_lateral_gr4_k)

lateral_SupplyRatio(ground_water_mm, ground_capacity_mm, param_lateral_sur_k)

lateral_SupplyPow(
  ground_water_mm,
  ground_capacity_mm,
  param_lateral_sup_k,
  param_lateral_sup_gamma
)
}
\arguments{
\item{ground_water_mm}{(mm/m2/TS) water volume in \code{groundLy}}

\item{ground_capacity_mm}{(mm/m2) water storage capacity in \code{groundLy}}

\item{param_lateral_gr4_k}{parameters}

\item{param_lateral_sur_k}{parameters for \code{\link[=lateral_SupplyRatio]{lateral_SupplyRatio()}}}

\item{param_lateral_sup_k, param_lateral_sup_gamma}{parameters for \code{\link[=lateral_SupplyPow]{lateral_SupplyPow()}}}
}
\value{
lateral_mm (mm/m2)
}
\description{
\strong{lateral flux}
}
