% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modells}
\alias{modells}
\alias{EDCHM_GR4J}
\alias{EDCHM_GR4J_full}
\alias{EDCHM_mini}
\alias{EDCHM_mini_full}
\alias{EDCHM_snow}
\alias{EDCHM_snow_full}
\title{modells build with EDCHM modulas}
\usage{
EDCHM_GR4J(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  S_,
  R_,
  X_1,
  X_2,
  X_3,
  X_4
)

EDCHM_GR4J_full(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  S_,
  R_,
  X_1,
  X_2,
  X_3,
  X_4
)

EDCHM_mini(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh
)

EDCHM_mini_full(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh
)

EDCHM_snow(
  n_time,
  n_spat,
  time_step_h,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  snow_ice_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_atmos_thr_Ts,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh,
  param_snow_fac_f,
  param_snow_fac_Tmelt
)

EDCHM_snow_full(
  n_time,
  n_spat,
  time_step_h,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  snow_ice_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_atmos_thr_Ts,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh,
  param_snow_fac_f,
  param_snow_fac_Tmelt
)
}
\arguments{
\item{n_time, n_spat}{number of time step and spatial unit}

\item{atmos_potentialEvatrans_mm}{(mm/m2/TS) \strong{potential / reference} evapotranspiration}

\item{atmos_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{S_, R_}{storage water S and R}

\item{X_1, X_2, X_3, X_4}{parameters in GR4J}

\item{ground_capacity_mm}{(mm/m2) water storage capacity in \code{groundLy}}

\item{ground_water_mm}{(mm/m2/TS) water volume in \code{groundLy}}

\item{land_impermeableFrac_1}{<0, 1> the maximum impermeable fraction when th soil is fully saturated}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{soil_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) \strong{potential} percolation}

\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{confluenLand_responseTime_TS, confluenGround_responseTime_TS, param_confluenLand_kel_k, }{parameters in modells}

\item{param_baseflow_grf_gamma}{<2, 7> exponential parameter for \code{\link[=baseflow_GR4Jfix]{baseflow_GR4Jfix()}}}

\item{param_evatrans_ubc_gamma}{<0.5, 2> parameter for \code{\link[=evatransActual_UBC]{evatransActual_UBC()}}}

\item{param_infilt_ubc_P0AGEN}{<0.1, 4> coefficient parameter for \code{\link[=infilt_UBC]{infilt_UBC()}}}

\item{param_percola_arn_k}{<0.1, 1> exponential parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{time_step_h}{<1, 24> (h) time step in hour}

\item{atmos_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{snow_ice_mm}{(mm/m2) water equivalent of \strong{ice} in snowpack}

\item{param_atmos_thr_Ts}{<-1, 3> (Cel) threshold air temperature that snow, parameter for \code{\link[=atmosSnow_ThresholdT]{atmosSnow_ThresholdT()}}}

\item{param_snow_fac_f}{<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for \code{\link[=snowMelt_Factor]{snowMelt_Factor()}}}

\item{param_snow_fac_Tmelt}{<0, 3> (Cel) snow melt temperature parameter for \code{\link[=snowMelt_Factor]{snowMelt_Factor()}}}
}
\value{
\itemize{
\item \code{EDCHM_xxxx}: stream flow in mm/TS
\item \code{EDCHM_xxxx_full}: list of variablen
}

stream flow in mm/TS
}
\description{
some example models with EDCHM modulas

some example models with EDCHM modulas
}
\section{\strong{EDCHM_GR4J} \insertCite{GR4J_Perrin_2003}{EDCHM}:}{
Total same like original GR4J
}

\section{\strong{EDCHM_GR4J_full} \insertCite{GR4J_Perrin_2003}{EDCHM}:}{
With all variable output
}

\section{\strong{EDCHM_mini}:}{
A model based on mini-structure with only six process:
\itemize{
\item evatransActual (evatransSoil)
\item infilt
\item percola
\item baseflow
\item confluenIUH (confluenLand)
\item confluenIUH (confluenGround)
}
}

\section{\strong{EDCHM_mini_full}:}{
With full Output
}

\section{\strong{EDCHM_snow}:}{
A model with \code{snowMelt} and \code{atmosSnow} based on mini-modell
}

\section{\strong{EDCHM_snow_full}:}{
With full Output
}

\references{
\insertAllCited{}

\insertAllCited{}
}
