% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{percola}
\alias{percola}
\alias{percola_GR4Jfix}
\alias{percola_GR4J}
\alias{percola_SupplyRatio}
\alias{percola_SupplyPow}
\alias{percola_MaxPow}
\alias{percola_ThreshPow}
\alias{percola_Arno}
\alias{percola_BevenWood}
\title{\strong{percolation}}
\usage{
percola_GR4Jfix(soil_water_mm, soil_capacity_mm, param_percola_grf_k)

percola_GR4J(soil_water_mm, soil_capacity_mm)

percola_SupplyRatio(soil_water_mm, param_percola_sur_k)

percola_SupplyPow(
  soil_water_mm,
  soil_capacity_mm,
  param_percola_sup_k,
  param_percola_sup_gamma
)

percola_MaxPow(
  soil_water_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  param_percola_map_gamma
)

percola_ThreshPow(
  soil_water_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  param_percola_thp_thresh,
  param_percola_thp_gamma
)

percola_Arno(
  soil_water_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  param_percola_arn_thresh,
  param_percola_arn_k
)

percola_BevenWood(
  soil_water_mm,
  soil_capacity_mm,
  soil_fieldCapacityPerc_1,
  soil_potentialPercola_mm
)
}
\arguments{
\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{param_percola_grf_k}{<0.01, 1> coefficient parameter for \code{\link[=percola_GR4Jfix]{percola_GR4Jfix()}}}

\item{param_percola_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=percola_SupplyRatio]{percola_SupplyRatio()}}}

\item{param_percola_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=percola_SupplyPow]{percola_SupplyPow()}}}

\item{param_percola_sup_gamma}{<0, 7> parameter for \code{\link[=percola_SupplyPow]{percola_SupplyPow()}}}

\item{soil_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) \strong{potential} percolation}

\item{param_percola_map_gamma}{<0.1, 5> exponential parameter for \code{\link[=percola_MaxPow]{percola_MaxPow()}}}

\item{param_percola_thp_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_thp_gamma}{<0.1, 5> exponential parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_arn_thresh}{<0.1, 0.9> coefficient parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{param_percola_arn_k}{<0.1, 1> exponential parameter for \code{\link[=percola_ThreshPow]{percola_ThreshPow()}}}

\item{soil_fieldCapacityPerc_1}{<0, 1> the relative ratio (\mjseqn{\theta_fc / \theta^*}) that the water content can drainage by gravity}
}
\value{
percola_mm (mm/m2)
}
\description{
\strong{percolation}
}
