% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestData.R
\docType{data}
\name{test_data}
\alias{test_data}
\alias{EDCHM_TestData}
\title{Test data for modelling}
\format{
An object of class \code{data.frame} with 180 rows and 14 columns.
}
\usage{
EDCHM_TestData
}
\description{
Precipitation, temperature and potential evapotranspiration
Test data contain three data precipitation (7 forms, start with \code{P}), temperature (5 forms, start with \code{T})
and potential evapotranspiration (2 forms, start with \code{E}),
each with 180 time steps. They can be used to test the model's run.
}
\keyword{datasets}
