% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{capirise}
\alias{capirise}
\alias{capirise_SupplyRatio}
\alias{capirise_SupplyPow}
\alias{capirise_AcceptRatio}
\alias{capirise_AcceptPow}
\title{\strong{capilarise}}
\usage{
capirise_SupplyRatio(
  ground_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_capirise_sur_k
)

capirise_SupplyPow(
  ground_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_capirise_sup_k,
  param_capirise_sup_gamma
)

capirise_AcceptRatio(
  ground_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_capirise_acr_k
)

capirise_AcceptPow(
  ground_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_capirise_acp_k,
  param_capirise_acp_gamma
)
}
\arguments{
\item{ground_water_mm}{(mm/m2/TS) water volume in \code{groundLy}}

\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{param_capirise_sur_k}{<0.01, 1> coefficient parameter for\code{\link[=capirise_SupplyRatio]{capirise_SupplyRatio()}}}

\item{param_capirise_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=capirise_SupplyPow]{capirise_SupplyPow()}}}

\item{param_capirise_sup_gamma}{<0.1, 5> exponential parameter for \code{\link[=capirise_SupplyPow]{capirise_SupplyPow()}}}

\item{param_capirise_acr_k}{<0.01, 1> coefficient parameter \code{\link[=capirise_AcceptRatio]{capirise_AcceptRatio()}}}

\item{param_capirise_acp_k}{<0.01, 1> coefficient parameter for \code{\link[=capirise_AcceptPow]{capirise_AcceptPow()}}}

\item{param_capirise_acp_gamma}{<0.1, 5> exponential parameter for \code{\link[=capirise_AcceptPow]{capirise_AcceptPow()}}}
}
\value{
ground_capillarise_mm (mm/m2/TS)
}
\description{
\strong{capilarise}
}
