% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{capirise}
\alias{capirise}
\alias{capirise_AcceptRatio}
\alias{capirise_AcceptPow}
\title{\strong{capilarise}}
\usage{
capirise_AcceptRatio(
  ground_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  soil_fieldCapacityPerc_1,
  param_capirise_acr_k
)

capirise_AcceptPow(
  ground_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  soil_fieldCapacityPerc_1,
  param_capirise_acp_k,
  param_capirise_acp_gamma
)
}
\arguments{
\item{ground_water_mm}{(mm/m2/TS) water volume in \code{groundLy}}

\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{soil_fieldCapacityPerc_1}{<0, 1> the relative ratio (\mjseqn{\theta_fc / \theta^*}) that the water content can drainage by gravity}

\item{param_capirise_acr_k}{<0.01, 1> coefficient parameter \code{\link[=capirise_AcceptRatio]{capirise_AcceptRatio()}}}

\item{param_capirise_acp_k}{<0.01, 1> coefficient parameter for \code{\link[=capirise_AcceptPow]{capirise_AcceptPow()}}}

\item{param_capirise_acp_gamma}{<0.01, 1> exponential parameter for \code{\link[=capirise_AcceptPow]{capirise_AcceptPow()}}}
}
\description{
\strong{capilarise}
}
