% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confluen}
\alias{confluen}
\alias{confluen_IUH}
\title{\strong{confluence}}
\usage{
confluen_IUH(confluen_inputWater_mm, confluen_iuh_1)
}
\arguments{
\item{confluen_inputWater_mm}{(mm/m2) input water volum in every routeline}

\item{confluen_iuh_1}{(vector of num, sume() = 1) the ratio in every timestep, can be calculated by \code{\link[=confluenIUH_GR4J1]{confluenIUH_GR4J1()}}, \code{\link[=confluenIUH_GR4J2]{confluenIUH_GR4J2()}}}
}
\value{
confluenced water (mm/m2)
}
\description{
\strong{confluence}
}
