% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confluenIUH}
\alias{confluenIUH}
\alias{confluenIUH_GR4J1}
\alias{confluenIUH_GR4J2}
\alias{confluenIUH_Clark}
\alias{confluenIUH_Kelly}
\alias{confluenIUH_Nash}
\alias{confluenIUH_NashKumar}
\title{create \strong{IUH} (Instant Unit Graphy)}
\usage{
confluenIUH_GR4J1(confluen_responseTime_TS)

confluenIUH_GR4J2(confluen_responseTime_TS)

confluenIUH_Clark(confluen_responseTime_TS)

confluenIUH_Kelly(confluen_responseTime_TS, param_confluen_kel_k)

confluenIUH_Nash(confluen_responseTime_TS, param_confluen_nas_n)

confluenIUH_NashKumar(confluen_responseTime_TS, param_confluen_nak_n)
}
\arguments{
\item{confluen_responseTime_TS}{(TS) response or concentration time in every routeline}

\item{param_confluen_kel_k}{<1, 4> parameter for\code{\link[=confluenIUH_Kelly]{confluenIUH_Kelly()}}}

\item{param_confluen_nas_n}{<1, 8> parameter for\code{\link[=confluenIUH_Nash]{confluenIUH_Nash()}}}

\item{param_confluen_nak_n}{<1, 8> parameter for\code{\link[=confluenIUH_NashKumar]{confluenIUH_NashKumar()}}}
}
\value{
IUH (list of num vector)
}
\description{
create \strong{IUH} (Instant Unit Graphy)
}
