% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evatransPotential}
\alias{evatransPotential}
\alias{evatransPotential_TurcWendling}
\alias{evatransPotential_Linacre}
\alias{evatransPotential_FAO56}
\title{\strong{potential evapotranspiration}}
\usage{
evatransPotential_TurcWendling(
  atmos_temperature_Cel,
  atmos_solarRadiat_MJ,
  time_step_h,
  param_evatrans_tur_k
)

evatransPotential_Linacre(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_relativeHumidity_1,
  land_latitude_Degree,
  land_elevation_m,
  land_albedo_1
)

evatransPotential_FAO56(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_vaporPress_hPa,
  atmos_saturatVaporPress_hPa,
  atmos_netRadiat_MJ,
  atmos_windSpeed2m_m_s,
  land_latitude_Degree,
  land_elevation_m
)
}
\arguments{
\item{atmos_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{atmos_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{time_step_h}{(1, 24 h) time step in hour}

\item{param_evatrans_tur_k}{parameters for \code{\link[=evatransPotential_TurcWendling]{evatransPotential_TurcWendling()}}}

\item{time_dayOfYear_}{(1, 366) the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)}

\item{atmos_relativeHumidity_1}{(0, 1) relative humidity}

\item{land_latitude_Degree}{(degree) average latitude}

\item{land_elevation_m}{(m) average elevation}

\item{land_albedo_1}{(0, 1) albedo of this region}

\item{atmos_vaporPress_hPa}{(hPa) actual vapour pressure, can be calculated by \code{\link[=atmos_VaporPress]{atmos_VaporPress()}}}

\item{atmos_saturatVaporPress_hPa}{(hPa) saturation vapour pressure at \code{atmos_temperature_Cel}, can be calculated by \code{\link[=atmos_SaturatVaporPress]{atmos_SaturatVaporPress()}}}

\item{atmos_netRadiat_MJ}{(MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation}

\item{atmos_windSpeed2m_m_s}{(m/s) wind speed at 2 m above ground surface}
}
\value{
potential evapotranspiration (mm/m2)
}
\description{
\strong{potential evapotranspiration}
}
