% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infilt}
\alias{infilt}
\alias{infilt_SupplyRatio}
\alias{infilt_SupplyPow}
\alias{infilt_AcceptRatio}
\alias{infilt_AcceptPow}
\alias{infilt_HBV}
\alias{infilt_GR4J}
\alias{infilt_SCS}
\alias{infilt_XAJ}
\title{\strong{infiltration}}
\usage{
infilt_SupplyRatio(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_sur_k
)

infilt_SupplyPow(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_sup_k,
  param_infilt_sup_gamma
)

infilt_AcceptRatio(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_acr_k
)

infilt_AcceptPow(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_acp_k,
  param_infilt_acp_gamma
)

infilt_HBV(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_hbv_beta
)

infilt_GR4J(land_water_mm, soil_water_mm, soil_capacity_mm)

infilt_SCS(land_water_mm, soil_water_mm, soil_capacity_mm, param_infilt_scs_CN)

infilt_XAJ(land_water_mm, soil_water_mm, soil_capacity_mm, param_infilt_xaj_B)
}
\arguments{
\item{land_water_mm}{(mm/m2) water volum in \code{landLy}, different than \code{land_interceptWater_mm}}

\item{soil_water_mm}{(mm/m2) water volum in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{param_infilt_sur_k}{parameters for \code{\link[=infilt_SupplyRatio]{infilt_SupplyRatio()}}}

\item{param_infilt_sup_k, param_infilt_sup_gamma}{parameters for \code{\link[=infilt_SupplyPow]{infilt_SupplyPow()}}}

\item{param_infilt_acr_k}{parameters for \code{\link[=infilt_AcceptRatio]{infilt_AcceptRatio()}}}

\item{param_infilt_acp_k, param_infilt_acp_gamma}{parameters for \code{\link[=infilt_AcceptPow]{infilt_AcceptPow()}}}

\item{param_infilt_hbv_beta}{parameters for \code{\link[=infilt_HBV]{infilt_HBV()}}}

\item{param_infilt_scs_CN}{parameters for \code{\link[=infilt_SCS]{infilt_SCS()}}}

\item{param_infilt_xaj_B}{parameters for \code{\link[=infilt_XAJ]{infilt_XAJ()}}}
}
\value{
infilt_mm (mm/m2)
}
\description{
\strong{infiltration}
}
