% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infilt}
\alias{infilt}
\alias{infilt_GR4J}
\alias{infilt_UBC}
\alias{infilt_SupplyRatio}
\alias{infilt_SupplyPow}
\alias{infilt_AcceptRatio}
\alias{infilt_AcceptPow}
\alias{infilt_HBV}
\alias{infilt_XAJ}
\alias{infilt_VIC}
\title{\strong{infiltration}}
\usage{
infilt_GR4J(land_water_mm, soil_water_mm, soil_capacity_mm)

infilt_UBC(
  land_water_mm,
  land_impermeableFrac_1,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_ubc_P0AGEN
)

infilt_SupplyRatio(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_sur_k
)

infilt_SupplyPow(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_sup_k,
  param_infilt_sup_gamma
)

infilt_AcceptRatio(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_acr_k
)

infilt_AcceptPow(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_acp_k,
  param_infilt_acp_gamma
)

infilt_HBV(
  land_water_mm,
  soil_water_mm,
  soil_capacity_mm,
  param_infilt_hbv_beta
)

infilt_XAJ(land_water_mm, soil_water_mm, soil_capacity_mm, param_infilt_xaj_B)

infilt_VIC(land_water_mm, soil_water_mm, soil_capacity_mm, param_infilt_vic_B)
}
\arguments{
\item{land_water_mm}{(mm/m2) \strong{pounded water} volume in \code{landLy} and there is no limit, different than \code{land_interceptWater_mm}}

\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{land_impermeableFrac_1}{<0, 1> the maximum impermeable fraction when th soil is fully saturated}

\item{param_infilt_ubc_P0AGEN}{<0.1, 4> coefficient parameter for \code{\link[=infilt_UBC]{infilt_UBC()}}}

\item{param_infilt_sur_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_SupplyRatio]{infilt_SupplyRatio()}}}

\item{param_infilt_sup_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_SupplyPow]{infilt_SupplyPow()}}}

\item{param_infilt_sup_gamma}{<0, 1> parameters for \code{\link[=infilt_SupplyPow]{infilt_SupplyPow()}}}

\item{param_infilt_acr_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_AcceptRatio]{infilt_AcceptRatio()}}}

\item{param_infilt_acp_k}{<0.01, 1> coefficient parameter for \code{\link[=infilt_AcceptPow]{infilt_AcceptPow()}}}

\item{param_infilt_acp_gamma}{<0.001, 5> parameters for \code{\link[=infilt_AcceptPow]{infilt_AcceptPow()}}}

\item{param_infilt_hbv_beta}{<0.001, 5> parameters for \code{\link[=infilt_HBV]{infilt_HBV()}}}

\item{param_infilt_xaj_B}{<0.01, 3> parameters for \code{\link[=infilt_XAJ]{infilt_XAJ()}}}

\item{param_infilt_vic_B}{<0.01, 3> parameters for \code{\link[=infilt_VIC]{infilt_VIC()}}}
}
\value{
infilt_mm (mm/m2)
}
\description{
\strong{infiltration}
}
