% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRange.R
\docType{data}
\name{ParamRange}
\alias{ParamRange}
\title{Parameter Range}
\format{
An object of class \code{data.frame} with 67 rows and 2 columns.
}
\usage{
ParamRange
}
\description{
Parameter range for calibration
}
\keyword{datasets}
