% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{percola}
\alias{percola}
\alias{percola_GR4J}
\alias{percola_SupplyRatio}
\alias{percola_SupplyPow}
\title{\strong{percolation}}
\usage{
percola_GR4J(soil_water_mm, soil_capacity_mm, param_percola_gr4_k)

percola_SupplyRatio(soil_water_mm, param_percola_sur_k)

percola_SupplyPow(
  soil_water_mm,
  soil_capacity_mm,
  param_percola_sup_k,
  param_percola_sup_gamma
)
}
\arguments{
\item{soil_water_mm}{(mm/m2) water volume in \code{soilLy}}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{param_percola_gr4_k}{parameters}

\item{param_percola_sur_k}{parameters for \code{\link[=percola_SupplyRatio]{percola_SupplyRatio()}}}

\item{param_percola_sup_k, param_percola_sup_gamma}{parameters for \code{\link[=percola_SupplyPow]{percola_SupplyPow()}}}
}
\value{
percola_mm (mm/m2)
}
\description{
\strong{percolation}
}
